+++ src/HHCP/NodeComparator.java
@@ -20,7 +20,7 @@ public class NodeComparator implements Comparator<Node>{
         }
         return 0;*/
     	//return Integer.compare(x.getFunction(), y.getFunction());
-    	return Integer.compare(x.getH(), y.getH());
+    	return Long.compare(x.getH(), y.getH());
     }
 
 }
